requirements.txt
alphashape==1.3.1
click==8.1.7
click-log==0.4.0
contourpy==1.1.1
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.47.2
importlib-metadata==7.0.1
importlib-resources==6.1.1
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.41.1
matplotlib==3.7.4
mordred==1.2.0
networkx==2.8.8
numba==0.58.1
numpy==1.24.4
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
pillow==10.2.0
point-cloud-utils==0.30.4
pynndescent==0.5.11
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.4
rdkit==2023.9.4
Rtree==1.2.0
scikit-learn==1.3.2
scipy==1.10.1
shapely==2.0.2
six==1.16.0
threadpoolctl==3.2.0
tqdm==4.66.1
trimesh==4.1.0
tzdata==2023.4
umap-learn==0.5.5
zipp==3.17.0
